Make LambdaClientFactory common to sink and processor
