+++ data-prepper-plugins/aws-lambda/src/main/java/org/opensearch/dataprepper/plugins/lambda/common/client/LambdaClientFactory.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package org.opensearch.dataprepper.plugins.lambda.sink;
+package org.opensearch.dataprepper.plugins.lambda.common.client;
 
 import org.opensearch.dataprepper.aws.api.AwsCredentialsOptions;
 import org.opensearch.dataprepper.aws.api.AwsCredentialsSupplier;
@@ -16,20 +16,21 @@ import software.amazon.awssdk.services.lambda.LambdaClient;
 public final class LambdaClientFactory {
     private LambdaClientFactory() { }
 
-    static LambdaClient createLambdaClient(final LambdaSinkConfig lambdaSinkConfig,
-                                           final AwsCredentialsSupplier awsCredentialsSupplier) {
-         final AwsCredentialsOptions awsCredentialsOptions = convertToCredentialsOptions(lambdaSinkConfig.getAwsAuthenticationOptions());
+    public static LambdaClient createLambdaClient(final AwsAuthenticationOptions awsAuthenticationOptions,
+                                                  final int maxConnectionRetries,
+                                                  final AwsCredentialsSupplier awsCredentialsSupplier) {
+         final AwsCredentialsOptions awsCredentialsOptions = convertToCredentialsOptions(awsAuthenticationOptions);
          final AwsCredentialsProvider awsCredentialsProvider = awsCredentialsSupplier.getProvider(awsCredentialsOptions);
 
             return LambdaClient.builder()
-                .region(lambdaSinkConfig.getAwsAuthenticationOptions().getAwsRegion())
+                .region(awsAuthenticationOptions.getAwsRegion())
                 .credentialsProvider(awsCredentialsProvider)
-                .overrideConfiguration(createOverrideConfiguration(lambdaSinkConfig)).build();
+                .overrideConfiguration(createOverrideConfiguration(maxConnectionRetries)).build();
 
     }
 
-    private static ClientOverrideConfiguration createOverrideConfiguration(final LambdaSinkConfig lambdaSinkConfig) {
-        final RetryPolicy retryPolicy = RetryPolicy.builder().numRetries(lambdaSinkConfig.getMaxConnectionRetries()).build();
+    private static ClientOverrideConfiguration createOverrideConfiguration(final int maxConnectionRetries) {
+        final RetryPolicy retryPolicy = RetryPolicy.builder().numRetries(maxConnectionRetries).build();
         return ClientOverrideConfiguration.builder()
                 .retryPolicy(retryPolicy)
                 .build();
